Add LGPL 2.0 LICENSE file
The README and source files clearly state that Bow is licensed under
the GNU Library General Public License, version 2.

To be explicit, and to allow GitHub to clearly identify the license,
I'm adding https://www.gnu.org/licenses/old-licenses/lgpl-2.0.txt
as the file LICENSE here.

This is merely in an attempt to clarify. It is not intended to
change the license or to relicense the files, only to make the original
license terms fully explicit. If there are any doubts or contradictions,
the comments in the original source code take precedence.